# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from "./orderItemInput";

export interface CreateOrderBody {
  customerName: string;
  customerEmail: string;
  customerPhone?: string;
  specialInstructions?: string;
  items: OrderItemInput[];
}