# lib/api-zod/src/generated/types/orderItemResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItemResponse {
  productName: string;
  quantity: number;
  unitPrice: string;
  lineTotal: string;
}