# lib/api-zod/src/generated/types/orderResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemResponse } from "./orderItemResponse";

export interface OrderResponse {
  id: number;
  customerName: string;
  customerEmail: string;
  subtotal: string;
  taxRate: string;
  taxAmount: string;
  total: string;
  status: string;
  createdAt: Date;
  items: OrderItemResponse[];
}